refactor: merge Runner class back into Codecept
The Runner extraction created a runtime object cycle: Codecept owned a
Runner, and Runner held a back-reference to Codecept for config, opts,
testFiles, requiringModules, loadTests, and the event.all.before/after
payload. Since Runner had one creation site, no independent state, and
its only collaborator was the object it pointed back to, the class
wasn't carrying its weight. Moving the methods back onto Codecept puts
behavior next to its data and eliminates the cycle by construction.

- lib/codecept.js: absorbs getSuites, run, runSuite, runTest, _execute
- lib/runner.js: deleted
- lib/index.js: Runner removed from public exports (unreleased API)
