+++ lib/codecept.js
@@ -10,15 +10,20 @@ import { createRequire } from 'module'
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
 
+import escapeRe from 'escape-string-regexp'
 import Helper from '@codeceptjs/helper'
-import Runner from './runner.js'
+import MochaFactory from './mocha/factory.js'
 import container from './container.js'
 import Config from './config.js'
+import event from './event.js'
+import recorder from './recorder.js'
+import output from './output.js'
 import runHook from './hooks.js'
 import ActorFactory from './actor.js'
 import { emptyFolder } from './utils.js'
 import { initCodeceptGlobals } from './globals.js'
 import store from './store.js'
+import { validateTypeScriptSetup, getTSNodeESMWarning } from './utils/loaderCheck.js'
 
 import storeListener from './listener/store.js'
 import stepsListener from './listener/steps.js'
@@ -100,7 +105,6 @@ class Codecept {
     await this.requireModules(this.requiringModules)
     // initializing listeners
     await container.create(this.config, this.opts)
-    this.runner = new Runner(this)
     await this.runHooks()
   }
 
@@ -263,7 +267,26 @@ class Codecept {
    * @returns {Array<{title: string, file: string, tags: string[], tests: Array<{title: string, uid: string, tags: string[], fullTitle: string}>}>}
    */
   getSuites(pattern) {
-    return this.runner.getSuites(pattern)
+    if (this.testFiles.length === 0) this.loadTests(pattern)
+
+    const tempMocha = MochaFactory.create(this.config.mocha || {}, this.opts || {})
+    tempMocha.files = this.testFiles
+    tempMocha.loadFiles()
+
+    const suites = []
+    for (const suite of tempMocha.suite.suites) {
+      suites.push({
+        ...suite.simplify(),
+        file: suite.file || '',
+        tests: suite.tests.map(test => ({
+          ...test.simplify(),
+          fullTitle: test.fullTitle(),
+        })),
+      })
+    }
+
+    tempMocha.unloadFiles()
+    return suites
   }
 
   /**
@@ -274,7 +297,7 @@ class Codecept {
    * @returns {Promise<void>}
    */
   async runSuite(suite) {
-    return this.runner.runSuite(suite)
+    return this.run(suite.file)
   }
 
   /**
@@ -285,7 +308,7 @@ class Codecept {
    * @returns {Promise<void>}
    */
   async runTest(test) {
-    return this.runner.runTest(test)
+    return this._execute({ grep: `^${escapeRe(test.fullTitle)}$` })
   }
 
   /**
@@ -295,7 +318,69 @@ class Codecept {
    * @returns {Promise<void>}
    */
   async run(test) {
-    return this.runner.run(test)
+    let files = this.testFiles
+
+    if (test) {
+      if (!fsPath.isAbsolute(test)) {
+        test = fsPath.join(store.codeceptDir, test)
+      }
+      const testBasename = fsPath.basename(test, '.js')
+      const testFeatureBasename = fsPath.basename(test, '.feature')
+      files = files.filter(t => {
+        return fsPath.basename(t, '.js') === testBasename || fsPath.basename(t, '.feature') === testFeatureBasename || t === test
+      })
+    }
+
+    return this._execute({ files })
+  }
+
+  async _execute({ files, grep } = {}) {
+    await container.started()
+
+    const tsValidation = validateTypeScriptSetup(this.testFiles, this.requiringModules || [])
+    if (tsValidation.hasError) {
+      output.error(tsValidation.message)
+      process.exit(1)
+    }
+
+    const tsWarning = getTSNodeESMWarning(this.requiringModules || [])
+    if (tsWarning) {
+      output.print(output.colors.yellow(tsWarning))
+    }
+
+    try {
+      const { loadTranslations } = await import('./mocha/gherkin.js')
+      await loadTranslations()
+    } catch (e) {
+      // Ignore if gherkin module not available
+    }
+
+    return new Promise((resolve, reject) => {
+      const mocha = container.mocha()
+      mocha.files = files || this.testFiles
+
+      if (grep) {
+        mocha.grep(grep)
+      }
+
+      const done = async (failures) => {
+        event.emit(event.all.result, container.result())
+        event.emit(event.all.after, this)
+        await recorder.promise()
+        if (failures) {
+          process.exitCode = 1
+        }
+        resolve()
+      }
+
+      try {
+        event.emit(event.all.before, this)
+        mocha.run(async (failures) => await done(failures))
+      } catch (e) {
+        output.error(e.stack)
+        reject(e)
+      }
+    })
   }
 
   /**
