+++ lib/index.js
@@ -24,7 +24,6 @@ import ai from './ai.js'
 import Workers from './workers.js'
 import Secret, { secret } from './secret.js'
 import Result from './result.js'
-import Runner from './runner.js'
 
 export default {
   /** @type {typeof CodeceptJS.Codecept} */
@@ -72,9 +71,7 @@ export default {
 
   /** @type {typeof Result} */
   Result,
-
-  Runner,
 }
 
 // Named exports for ESM compatibility
-export { codecept, codecept as Codecept, output, container, event, recorder, config, actor, helper, helper as Helper, pause, within, dataTable, dataTableArgument, store, locator, heal, ai, Workers, Secret, secret, Result, Runner }
+export { codecept, codecept as Codecept, output, container, event, recorder, config, actor, helper, helper as Helper, pause, within, dataTable, dataTableArgument, store, locator, heal, ai, Workers, Secret, secret, Result }
